+++ src/demo/unicodeblocks.c
@@ -119,7 +119,7 @@ draw_block(struct ncplane* nn, uint32_t blockstart){
 }
 
 int unicodeblocks_demo(struct notcurses* nc){
-  int maxx, maxy;
+  unsigned maxx, maxy;
   struct ncplane* n = notcurses_stddim_yx(nc, &maxy, &maxx);
   // some blocks are good for the printing, some less so. some are only
   // marginally covered by mainstream fonts, some not at all. we explicitly