+++ NEWS.md
@@ -12,6 +12,8 @@ rearrangements of Notcurses.
     deprecated functionality, ABI3 ought require small changes, if any.
 
 * 2.4.9 (not yet released)
+  * `qprefix()`, `bprefix()`, and `iprefix()` have been renamed
+    `ncqprefix()`, `ncbprefix()`, and `nciprefix()`, respectively.
   * `notcurses_mice_enable()` and `notcurses_mouse_disable()` replace
     `notcurses_mouse_enable()` and `notcurses_mouse_disable()`, which
     have been deprecated, and will be removed in ABI3.